[FIX] Proper error handling for modules with errors.
Withouth this, when loading odoo, while a module, installed or not, contains any error, Odoo will simply
stop, withouth ginving the least hint or message what is wrong. Nothing will be logged whatsoever.
With this fix, not only will a proper error and traceback be logged, but odoo loading will continue,
making it possible to work, even if some not installed modules contain errors.